refactor(terminal): split file link detection
